Remove support for multiple stashes, move to single stash

+++ src/lib.rs
@@ -3,35 +3,62 @@ use aes_gcm::{
     Aes256Gcm,
 };
 use std::io::{self, Read, Seek, Write};
-use std::{env, fs, path::Path};
+use std::{
+    env, fs,
+    path::{Path, PathBuf},
+};
+#[cfg(test)]
+use tempfile::TempDir;
 
-//  Create a new stash directory at `label` in current directory.
-pub fn init_stash(path: &str, label: &str) -> io::Result<()> {
-    let new_stash = format!("{}/{}", path, label);
-    fs::create_dir(&new_stash)?;
+//  Specify `~/.stash` as default path; use temp directory when testing
+//  TODO: add Result as return value
+fn get_stash_path() -> PathBuf {
+    #[cfg(not(test))]
+    {
+        let home = match env::var("HOME") {
+            Ok(path) => PathBuf::from(path),
+            Err(_) => panic!("Failed to get `HOME` environment variable"),
+        };
+        home.join(".stash")
+    }
+    #[cfg(test)]
+    {
+        let temp_dir = TempDir::new().unwrap();
+        temp_dir.path().join("test_stash")
+    }
+}
 
-    //  Create a `tar.gz` of new, empty directory
-    let tarball = format!("{}/contents", new_stash);
-    create_tarball(&new_stash, &tarball)?;
-    let contents = Path::new(&tarball);
+//  Create a new stash in user's home directory
+pub fn init() -> io::Result<()> {
+    let stash_path = get_stash_path();
+    //let contents = stash_path.join("contents");
 
-    //  Generate key and perform encryption
-    let key = Aes256Gcm::generate_key(OsRng);
-    encrypt(contents, &key)?;
-    //encrypt_a_copy(contents, &key)?;
+    if stash_path.exists() {
+        return Err(io::Error::new(
+            io::ErrorKind::AlreadyExists,
+            "Stash already exists.",
+        ));
+    }
+
+    fs::create_dir_all(stash_path.to_str().unwrap())?;
+    //create_tarball(&contents)?;
+
+    //let stash_key = Aes256Gcm::generate_key(OsRng);
+    //encrypt(&contents, &stash_key)?;
 
     //  TODO: store key on keyring
-    //  TODO: reset authentication timeout (eventually)
 
     Ok(())
 }
 
-//  TODO: slice of strings or space-separated filepaths as `label`
-fn create_tarball(path: &str, label: &str) -> Result<(), io::Error> {
+//  Create a `.tar.gz` archive
+#[allow(dead_code)]
+fn create_tarball(archive_path: &Path) -> Result<(), io::Error> {
+    let stash_path = get_stash_path();
     let output = std::process::Command::new("tar")
         .arg("czf")
-        .arg(label)
-        .arg(path)
+        .arg(archive_path)
+        .arg(stash_path)
         .output()?;
 
     if !output.status.success() {
@@ -46,8 +73,7 @@ fn create_tarball(path: &str, label: &str) -> Result<(), io::Error> {
     Ok(())
 }
 
-#[allow(dead_code)]
-//  Encrypt a specified file in place.
+//  Encrypt a specified file in place
 fn encrypt(path: &Path, key: &[u8]) -> io::Result<()> {
     let mut file = fs::OpenOptions::new().read(true).write(true).open(path)?;
     let mut buffer = Vec::new();
@@ -66,12 +92,15 @@ fn encrypt(path: &Path, key: &[u8]) -> io::Result<()> {
     Ok(())
 }
 
-#[allow(dead_code)]
-//  Create an encrypted copy of a specified file.
+//  Create an encrypted copy of a specified file
 fn encrypt_a_copy(path: &Path, key: &[u8]) -> io::Result<()> {
-    //  TODO: specify src/dst somehow
+    let stash_path = get_stash_path();
+    if !stash_path.exists() {
+        return Err(io::Error::new(io::ErrorKind::NotFound, "No stash found"));
+    }
+    let output_path = stash_path.join(path.file_name().unwrap());
+
     let mut input_file = fs::File::open(path)?;
-    let output_path = path.with_extension("enc");
     let mut output_file = fs::File::create(output_path)?;
 
     let mut input_buffer = Vec::new();
@@ -86,45 +115,80 @@ fn encrypt_a_copy(path: &Path, key: &[u8]) -> io::Result<()> {
     Ok(())
 }
 
-//  List all files in stash directory at `label` in current directory.
-pub fn list_stash(label: &str) -> io::Result<String> {
-    let stash_path = env::current_dir()?.join(label);
+//  List all files in stash directory
+pub fn list() -> io::Result<String> {
+    let stash_path = get_stash_path();
+    if !stash_path.exists() {
+        return Err(io::Error::new(io::ErrorKind::NotFound, "No stash found"));
+    }
+
+    //  TODO: this will need to decrypt the `contents` file
+    //  TODO: it will then need to run some form of `tar -t` instead
     let ls = std::process::Command::new("ls")
-        .arg(&stash_path)
+        .arg(stash_path.to_str().unwrap())
         .output()
         .expect("Failed to execute ls command")
         .stdout;
 
     let contents = String::from_utf8_lossy(&ls).trim().to_string();
 
+    //  TODO: re-encrypt
+
     Ok(contents)
 }
 
-//  Move `file` into stash at `label`
-pub fn move_file(file: &str, label: &str) -> io::Result<()> {
-    let file_path = Path::new(file);
-    let stash_path = Path::new(label).join(file_path.file_name().unwrap());
-    fs::rename(file_path, stash_path)?;
+//  Add `file` to stash
+pub fn add(file: &str) -> io::Result<()> {
+    let stash_path = get_stash_path();
+    if !stash_path.exists() {
+        return Err(io::Error::new(io::ErrorKind::NotFound, "No stash found"));
+    }
+
+    let src_path = Path::new(file);
+    let dst_path = stash_path.join(src_path.file_name().unwrap());
+    fs::rename(src_path, &dst_path).unwrap();
+
+    let file_key = Aes256Gcm::generate_key(OsRng);
+    encrypt(&dst_path, &file_key)?;
+
+    //  TODO: decrypt `contents` file (tarball) using `stash_key`
+    //  TODO: unpack decrypted tarball
+    //  TODO: create new tarball of stash contents
+    //  TODO: re-encrypt `contents` using `stash_key`
 
     Ok(())
 }
 
-//  Copy `file` into stash at `label`
-pub fn copy_file(file: &str, label: &str) -> io::Result<()> {
-    let file_path = Path::new(file);
-    let stash_path = Path::new(label).join(file_path.file_name().unwrap());
-    fs::copy(file_path, stash_path)?;
+//  Copy `file` into stash
+pub fn copy(file: &str) -> io::Result<()> {
+    let stash_path = get_stash_path();
+    if !stash_path.exists() {
+        return Err(io::Error::new(io::ErrorKind::NotFound, "No stash found"));
+    }
+
+    let src_path = Path::new(file);
+    let dst_path = stash_path.join(src_path.file_name().unwrap());
+    fs::copy(src_path, &dst_path).unwrap();
+
+    let file_key = Aes256Gcm::generate_key(OsRng);
+    encrypt_a_copy(&dst_path, &file_key)?;
+
+    //  TODO: decrypt `contents` file (tarball) using `stash_key`
+    //  TODO: unpack decrypted tarball
+    //  TODO: create new tarball of stash contents
+    //  TODO: re-encrypt `contents` using `stash_key`
 
     Ok(())
 }
 
-//  Move `file` from stash `label` to `file` in current directory.
-pub fn grab_file(file: &str, label: &str) -> io::Result<()> {
-    let src = format!("{}/{}", label, file);
-    let src_path = Path::new(&src);
-
-    let dst = format!("{}/{}", env::current_dir()?.display(), file);
-    let dst_path = Path::new(&dst);
+//  Move `file` from stash into current directory
+pub fn grab(file: &str) -> io::Result<()> {
+    let stash_path = get_stash_path();
+    if !stash_path.exists() {
+        return Err(io::Error::new(io::ErrorKind::NotFound, "No stash found"));
+    }
+    let src_path = stash_path.join(file);
+    let dst_path = env::current_dir()?.join(file);
 
     fs::rename(src_path, dst_path)?;
